fixed create-bookshelf error and made ui updates, overall ui/ux still work in progress
